added viewing attendees
