+++ screens/postDetails.js
@@ -1,16 +1,21 @@
-import React, { useState, useEffect } from "react";
-import { Alert, Text, View, TouchableOpacity } from "react-native";
+import React, {useContext, useState, useEffect } from "react";
+import { Alert,ScrollView, Modal,Text, View, TouchableOpacity,ActivityIndicator } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import moment from "moment";
-import { Button } from "react-native-paper";
 
+import { modalStyles } from '../styles/modal';
 import { postDetailsStyles } from "../styles/postDetails";
 import { globalStyles } from "../styles/global";
+import { UserContext } from "../util/GlobalStateManager";
 
 export default function PostScreen({ route, navigation }) {
   const poster = route.params.item.studentemail;
+  const { user } = useContext(UserContext);
   const [data, setData] = useState([]);
   const [isLoading, setLoading] = useState(true);
+  const [attendeeVisible, setAttendeeVisible] =useState(false);
+  const [attendees, setAttendees] = useState([]);
+  const [isModalVisible, setModalVisible] =useState(false);
 
   const getStudent = async () => {
     try {
@@ -28,7 +33,40 @@ export default function PostScreen({ route, navigation }) {
 
   useEffect(() => {
     getStudent();
+    getAttendees();
   }, []);
+
+  const getAttendees = async () => {
+    try {
+      const response = await fetch('https://knight-bites.herokuapp.com/attendees/'+ route.params.item.id); //postid
+      const json = await response.json();
+      setAttendees(json);
+    } catch (error) {
+      console.error(error);
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  const addAttendee = async () =>{
+    await fetch('https://knight-bites.herokuapp.com/attendees/' + route.params.item.id, {
+      method: 'POST',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        'postid': route.params.item.id,
+        'studentemail': user
+      })
+    })
+    .then((responseJson) => {
+      console.log('response object:' , JSON.stringify(responseJson))
+    })
+    .catch((error) => {
+      console.error(error);
+    })
+  };
 
   const iconName = data.icon;
 
@@ -71,15 +109,54 @@ export default function PostScreen({ route, navigation }) {
             {"\n"}
           </Text>
         </View>
-        <Button
+        <TouchableOpacity
           onPress={() => {
             Alert.alert("Signed up for event!");
           }}
           style={postDetailsStyles.signupButton}
         >
           <Text style={postDetailsStyles.signupButtonText}>Attend</Text>
-        </Button>
-      </View>
+        </TouchableOpacity>
+
+        <View style={postDetailsStyles.signupButtonText}>
+          <TouchableOpacity onPress={() => {setAttendeeVisible(true); setModalVisible(true)}}>
+            <Text>{'\n\n'}Who's going?</Text>
+          </TouchableOpacity>
+        </View>
+      
+      {/* To show the names of attendees for each event */}
+      <Modal
+          animationType="slide"
+          transparent={true}
+          visible={isModalVisible}
+          onRequestClose={() => {
+            Alert.alert("No changes made");
+            setModalVisible(!isModalVisible);
+            //setButtonVisible(true);
+          }}>
+          {attendeeVisible ?
+          <View style={modalStyles.centeredView}>
+            <View style={modalStyles.modalView}>
+              <Text style={postDetailsStyles.title}>Who's going?{'\n\n'}</Text>
+            {isLoading ? <ActivityIndicator />:(
+                <ScrollView >
+                {
+                  attendees.map((item, index) => {
+                    return (
+                      <Text key={index} style= {postDetailsStyles.body}>{item.firstname} {item.lastname}</Text>
+                    )
+                  })
+                }
+                </ScrollView>
+            )}
+            <TouchableOpacity onPress={()=>{setModalVisible(false); setAttendeeVisible(false)}}>
+              <Text style= {{fontSize:30}}>OK</Text>
+            </TouchableOpacity>
+            </View>
+          </View>
+          :null}
+        </Modal>
+        </View>
     </View>
   );
 }
